[Refactor] Updated User DTOs for improved User Registration
